refactor: extract external_data_concern
